# schemas/execution-gateway/local-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bean Execution Gateway API",
    "version": "0.1.0",
    "description": "Guardrailed API for route decisions, outcome records, and ledger summaries. No external execution.",
    "summary": "Hosted-demo mode must use public/synthetic inputs only and must not persist request bodies."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Localhost deployment"
    },
    {
      "url": "https://bean-execution-gateway-poc.onrender.com",
      "description": "Public hosted demo on Render Free"
    }
  ],
...
  "paths": {
...
    "/v0/route": {
      "post": {
        "operationId": "createExecutionRouteV0",
        "summary": "Find a policy-gated execution path for public or synthetic input",
        "description": "Returns a route decision only. It does not execute external actions, spend money, call suppliers, or persist request bodies in hosted-demo mode.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionGatewayRequestV0"
              },
              "examples": {
                "allowed": {
                  "summary": "Public issue triage",
                  "value": {
                    "outcome": {
                      "goal": "Triage a public GitHub issue and produce a local next-step packet.",
                      "task_type": "issue_triage",
                      "desired_artifact": "triage_packet",
                      "success_criteria": [
                        "local artifact",
                        "verifier plan",
                        "no public comment"
                      ]
                    },
                    "context_refs": [
                      "https://github.com/example/project/issues/101"
                    ],
                    "policy": {
                      "mode": "free_only"
                    }
                  }
                },
                "blocked": {
                  "summary": "Paid public write is blocked",
                  "value": {
                    "outcome": {
                      "goal": "Use a paid API and post a comment on the public issue.",
                      "task_type": "agent_task_triage",
                      "desired_artifact": "blocked_packet"
                    },
                    "context_refs": [
                      "https://github.com/example/project/issues/102"
                    ],
                    "policy": {
                      "mode": "free_only"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ExecutionGatewayResponseV0 route decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionGatewayResponseV0"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or hosted-demo private/secret input rejection",
            "content": {
              "application/json": {
                "examples": {
                  "hostedReject": {
                    "value": {
                      "api_version": "v0",
                      "ok": false,
                      "hosted_demo": true,
                      "error": "hosted_demo_rejects_private_work_or_secret_like_context",
                      "warning": "Do not paste private, work, company, customer, secret, credential, local file, internal, or regulated context into the public demo."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v0/dispatch": {
      "post": {
        "operationId": "createExecutionDispatchV0",
        "summary": "Prove external dispatch is disabled in V0",
        "description": "Always returns 403 in V0. Dispatch, supplier calls, paid APIs, account mutation, and public writes are intentionally unavailable.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionDispatchRequestV0"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Dispatch is disabled in V0",
            "content": {
              "application/json": {
                "examples": {
                  "disabled": {
                    "value": {
                      "api_version": "v0",
                      "ok": false,
                      "error": "dispatch_disabled_in_v0",
                      "spend_usd": 0,
                      "external_writes": 0,
                      "external_executions": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/route": {
      "post": {
        "operationId": "createExecutionRouteV0",
        "summary": "Find a policy-gated execution path for public or synthetic input",
        "description": "Returns a route decision only. It does not execute external actions, spend money, call suppliers, or persist request bodies in hosted-demo mode.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionGatewayRequestV0"
              },
              "examples": {
                "allowed": {
                  "summary": "Public issue triage",
                  "value": {
                    "outcome": {
                      "goal": "Triage a public GitHub issue and produce a local next-step packet.",
                      "task_type": "issue_triage",
                      "desired_artifact": "triage_packet",
                      "success_criteria": [
                        "local artifact",
                        "verifier plan",
                        "no public comment"
                      ]
                    },
                    "context_refs": [
                      "https://github.com/example/project/issues/101"
                    ],
                    "policy": {
                      "mode": "free_only"
                    }
                  }
                },
                "blocked": {
                  "summary": "Paid public write is blocked",
                  "value": {
                    "outcome": {
                      "goal": "Use a paid API and post a comment on the public issue.",
                      "task_type": "agent_task_triage",
                      "desired_artifact": "blocked_packet"
                    },
                    "context_refs": [
                      "https://github.com/example/project/issues/102"
                    ],
                    "policy": {
                      "mode": "free_only"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ExecutionGatewayResponseV0 route decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionGatewayResponseV0"
                }
              }
            }
          },
          "400": {
            "description": "Invalid JSON or hosted-demo private/secret input rejection",
            "content": {
              "application/json": {
                "examples": {
                  "hostedReject": {
                    "value": {
                      "api_version": "v0",
                      "ok": false,
                      "hosted_demo": true,
                      "error": "hosted_demo_rejects_private_work_or_secret_like_context",
                      "warning": "Do not paste private, work, company, customer, secret, credential, local file, internal, or regulated context into the public demo."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dispatch": {
      "post": {
        "operationId": "createExecutionDispatchV0",
        "summary": "Prove external dispatch is disabled in V0",
        "description": "Always returns 403 in V0. Dispatch, supplier calls, paid APIs, account mutation, and public writes are intentionally unavailable.",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutionDispatchRequestV0"
              }
            }
          }
        },
        "responses": {
          "403": {
            "description": "Dispatch is disabled in V0",
            "content": {
              "application/json": {
                "examples": {
                  "disabled": {
                    "value": {
                      "api_version": "v0",
                      "ok": false,
                      "error": "dispatch_disabled_in_v0",
                      "spend_usd": 0,
                      "external_writes": 0,
                      "external_executions": 0
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ExecutionGatewayRequestV0": {
        "type": "object",
        "required": [
          "outcome"
        ],
        "properties": {
          "outcome": {
            "type": "object",
            "required": [
              "goal"
            ]
          },
          "context_refs": {
            "type": "array"
          },
          "policy": {
            "type": "object"
          }
        },
        "additionalProperties": true
      },
      "ExecutionGatewayResponseV0": {
        "type": "object",
        "required": [
          "route_run_id",
          "request_id",
          "decision",
          "cost",
          "stop_conditions",
          "sanitization_report"
        ],
        "properties": {
          "api_version": {
            "type": "string",
            "const": "v0"
          },
          "route_run_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "decision": {
            "type": "object"
          },
          "cost": {
            "type": "object"
          },
          "stop_conditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sanitization_report": {
            "type": "object"
          },
          "hosted_demo": {
            "type": "boolean"
          }
        },
        "additionalProperties": true
      },
      "ExecutionDispatchRequestV0": {
        "type": "object",
        "properties": {
          "route_run_id": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "desired_outcome": {
            "type": "object"
          },
          "max_budget_usd": {
            "type": "number",
            "const": 0
          }
        },
        "additionalProperties": true
      },
...
    }
  }
}